--- app/src/main/res/layout/view_card_swap.xml
@@ -1,97 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@color/lawrence">
-
-    <ImageView
-        android:id="@+id/iconCoin"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="12dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:src="@drawable/ethereum" />
-
-    <TextView
-        android:id="@+id/titleTextView"
-        style="@style/Body"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:ellipsize="end"
-        android:lines="1"
-        android:textColor="@color/leah"
-        app:layout_constrainedWidth="true"
-        app:layout_constraintBottom_toBottomOf="@id/iconCoin"
-        app:layout_constraintEnd_toStartOf="@+id/selectedToken"
-        app:layout_constraintHorizontal_bias="0"
-        app:layout_constraintHorizontal_chainStyle="packed"
-        app:layout_constraintStart_toEndOf="@id/iconCoin"
-        app:layout_constraintTop_toTopOf="@id/iconCoin"
-        tools:text="You Pay" />
-
-    <TextView
-        android:id="@+id/selectedToken"
-        style="@style/Subhead1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="13dp"
-        android:layout_marginEnd="16dp"
-        android:drawableEnd="@drawable/ic_down_arrow_20"
-        android:drawablePadding="8dp"
-        android:textColor="@color/jacob"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:text="@string/Swap_TokenSelectorTitle" />
-
-    <View
-        android:id="@+id/background"
-        android:layout_width="0dp"
-        android:layout_height="84dp"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="12dp"
-        android:layout_marginEnd="8dp"
-        android:background="@drawable/swap_input_background"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/titleTextView" />
-
-    <io.horizontalsystems.bankwallet.ui.extensions.AmountInputView
-        android:id="@+id/amountInput"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        app:layout_constraintEnd_toEndOf="@id/background"
-        app:layout_constraintStart_toStartOf="@id/background"
-        app:layout_constraintTop_toTopOf="@id/background" />
-
-    <TextView
-        android:id="@+id/balanceTitle"
-        style="@style/Subhead2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="12dp"
-        android:layout_marginBottom="12dp"
-        android:text="@string/Swap_Balance"
-        android:textColor="@color/grey"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/background" />
-
-    <TextView
-        android:id="@+id/balanceValue"
-        style="@style/Subhead2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:text="@string/Swap_AmountHint"
-        android:textColor="@color/grey"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="@id/balanceTitle"
-        tools:text="0.12345678 XRP" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
